+++ http4k-format/core/src/main/kotlin/org/http4k/events/AutoJsonEvents.kt
@@ -1,9 +1,12 @@
 package org.http4k.events
 
-import org.http4k.format.AutoMarshallingJson
+import org.http4k.format.AutoMarshalling
 
-object AutoJsonEvents {
-    operator fun invoke(json: AutoMarshallingJson, print: (String) -> Unit = ::println) = object : Events {
+@Deprecated("Replace with AutoMarshallingEvents", ReplaceWith("AutoMarshallingEvents"))
+typealias AutoJsonEvents = AutoMarshallingEvents
+
+object AutoMarshallingEvents {
+    operator fun invoke(json: AutoMarshalling, print: (String) -> Unit = ::println) = object : Events {
         override fun invoke(p1: Event) = print(json.asFormatString(p1))
     }
 }