+++ http4k-format/core/src/test/kotlin/org/http4k/format/AutoMarshallingEventsContract.kt
@@ -4,7 +4,7 @@ import org.http4k.core.ContentType.Companion.APPLICATION_JSON
 import org.http4k.core.Response
 import org.http4k.core.Status.Companion.OK
 import org.http4k.core.with
-import org.http4k.events.AutoJsonEvents
+import org.http4k.events.AutoMarshallingEvents
 import org.http4k.events.EventFilters
 import org.http4k.events.MyEvent
 import org.http4k.events.plus
@@ -12,15 +12,15 @@ import org.http4k.events.then
 import org.http4k.filter.TraceId
 import org.http4k.filter.ZipkinTraces
 import org.http4k.lens.Header.CONTENT_TYPE
+import org.http4k.testing.ApprovalTest
 import org.http4k.testing.Approver
-import org.http4k.testing.JsonApprovalTest
 import org.http4k.util.FixedClock
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import java.io.StringWriter
 
-@ExtendWith(JsonApprovalTest::class)
-abstract class AutoJsonEventsContract(private val j: AutoMarshallingJson) {
+@ExtendWith(ApprovalTest::class)
+abstract class AutoMarshallingEventsContract(private val j: AutoMarshallingJson) {
 
     @Test
     fun `event serialises to json`(approver: Approver) {
@@ -35,7 +35,7 @@ abstract class AutoJsonEventsContract(private val j: AutoMarshallingJson) {
 
         val pipeline = EventFilters.AddZipkinTraces()
             .then(EventFilters.AddTimestamp(FixedClock))
-            .then(AutoJsonEvents(j, w::write))
+            .then(AutoMarshallingEvents(j, w::write))
 
         pipeline(final)