+++ http4k-format/klaxon/src/test/kotlin/org/http4k/format/klaxonJsonTests.kt
@@ -53,4 +53,4 @@ class KlaxonAutoTest : AutoMarshallingJsonContract(Klaxon) {
 
 }
 
-class KlaxonAutoEventsTest : AutoJsonEventsContract(Klaxon)
+class KlaxonAutoEventsTest : AutoMarshallingEventsContract(Klaxon)